teleport: organize user export fixtures
