+++ packages/teleport/__tests__/exporter/raw.spec.js
@@ -5,7 +5,7 @@ import {expect} from 'chai';
 import {exportUserRows} from '../../lib/exporter/raw.js';
 import {runSqlFile, useDatabase} from '../_utils/test-db.js';
 import {createKnexProxy} from '../../lib/shared/db.js';
-import {rawUserExport} from '../fixtures/raw-user-export.js';
+import {rawUserExport} from '../fixtures/user-export.js';
 
 const __dirname = path.dirname(fileURLToPath(import.meta.url));
