+++ packages/teleport/__tests__/importer/raw.spec.js
@@ -6,7 +6,7 @@ import {remapRawExportIds} from '../../lib/shared/remap-id.js';
 import {importUserRows} from '../../lib/importer/raw.js';
 import {runSqlFile, useDatabase} from '../_utils/test-db.js';
 import {createKnexProxy} from '../../lib/shared/db.js';
-import {rawUserExport} from '../fixtures/raw-user-export.js';
+import {rawUserExport} from '../fixtures/user-export.js';
 import {exportUserRows} from '../../lib/exporter/raw.js';
 
 const __dirname = path.dirname(fileURLToPath(import.meta.url));
